[REFATORAÇÃO] - Separação de métodos nos testes da carteira 17 variação 043 brasil e organização do teste na pasta do banco
